generator/pyproject.toml
[project]
name = "noteblock-generator"
version = "0.3.0"
description = "Generate music in Minecraft noteblocks"
authors = [{ name = "Felix Fourcolor", email = "felix.fourcolor@gmail.com" }]
license = { text = "MIT" }
requires-python = ">=3.10, <3.15"
keywords = ["minecraft", "noteblock"]
dependencies = [
    "amulet-core>=1.9.32",
    "typer>=0.20.0",
    "msgspec>=0.20.0",
    "watchfiles>=1.1.1",
]

[project.urls]
Homepage = "https://www.youtube.com/@felixfourcolor"
Repository = "https://github.com/FelixFourcolor/noteblock-generator"

[project.scripts]
nbg = "noteblock_generator.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["noteblock_generator"]

[tool.ruff]
target-version = "py310"
show-fixes = true
extend-safe-fixes = ["ALL"]

[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "standard"

[dependency-groups]
dev = ["pytest>=9.0.2"]
